requirements.txt
streamlit==1.28.1
pymongo==4.5.0
pandas==2.1.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
openpyxl==3.1.2